api/migrations/models/6_20250421012902_add_unique_constraint_to_grants.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        CREATE UNIQUE INDEX IF NOT EXISTS "uid_grants_year_fc9f6e" ON "grants" ("year", "office_id", "sn");"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        DROP INDEX IF EXISTS "uid_grants_year_fc9f6e";"""
